Adjusted projects api
